requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.1
python-multipart==0.0.9
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0 